requirements.txt
django
python-dotenv
psycopg2
Pillow
requests
langchain-gigachat